Move back JsonToRecord to initial package
