+++ common-stream-io/stream-json/src/main/java/org/talend/components/common/stream/input/json/JsonToRecord.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.common.stream.input.json.jsonToRecord;
+package org.talend.components.common.stream.input.json;
 
 import org.talend.sdk.component.api.record.Record;
 import org.talend.sdk.component.api.record.Schema;
